[nrunner] introduce a mechanism to serialize sets in config
This is a pre-work to passing all the config options to a runnable. This
handles the nrunner serialization/deserialization process involved in
the config.

It introduces a mechanism that converts sets into lists inside a
dictionary. The dictionary key `__encoded_set__` is used in the decoding
process to convert the option back to a set.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>