check.py/setup.py: Implement --select / --skip arguments
With --select and --skip we have a bigger flexibility choosing
the tests we want to run.

By default, with no argument, it'll run all tests

  python3 selftests/check.py

To run all the test except a given one, we can use --skip

  python3 selftests/check.py --skip static-checks

To run only one or several selected tests, use --select

  python3 selftests/check.py --select static-checks,unit

If we also want to avoid running the tests of one of the plugins:

  python3 selftests/check.py --skip nrunner-interface,static-check \
                             --disable-plugin-checks golang,robot

Add also a hidden argument --dict-tests that stores a dict with
the list of tests provided by --select and --skip.

Sync setup.py with selftests/check.py after introducing --select
and --skip.

Split the suites. There was a suite named "check", that grouped all
the tests except the ones from job-api and nrunner-interface.
Now we have suites that follow the name of the test groups:
jobs, unit, functional, static-checks and optional-plugins.

Update cirrus.yml,.github/workflows/weekly.yml and python-avocado.spec
to work with new command options in selftests/check.py

Signed-off-by: Ana Guerrero Lopez <anguerre@redhat.com>